Fix: Address review comments on PR #4
