+++ VideoGameDatabase/VideoGameDatabase/App/AppDelegate.swift
@@ -7,14 +7,13 @@
 
 import UIKit
 import Firebase
+
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
-
-
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
-        FirebaseApp.configure()
+        configFirebase()
         return true
     }
 
@@ -31,7 +30,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
         // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
     }
-
-
+    
+    private func configFirebase() {
+            FirebaseApp.configure()
+        }
 }
