+++ VideoGameDatabase/VideoGameDatabase/Models/Developer.swift
@@ -1,0 +1,12 @@
+//
+//  Developer.swift
+//  VideoGameDatabase
+//
+//  Created by macbook on 12/8/25.
+//
+import Foundation
+
+struct Developer: Codable {
+    let id: Int
+    let name: String
+}
