+++ VideoGameDatabase/VideoGameDatabase/Models/GamePublisher.swift
@@ -1,0 +1,13 @@
+//
+//  GamePublisher.swift
+//  VideoGameDatabase
+//
+//  Created by macbook on 12/8/25.
+//
+
+import Foundation
+
+struct GamePublisher: Codable {
+    let id: Int
+    let name: String
+}
