+++ VideoGameDatabase/VideoGameDatabase/Models/Genre.swift
@@ -1,0 +1,12 @@
+//
+//  Genre.swift
+//  VideoGameDatabase
+//
+//  Created by macbook on 12/8/25.
+//
+import Foundation
+
+struct Genre: Codable {
+    let id: Int
+    let name: String
+}
