+++ VideoGameDatabase/VideoGameDatabase/Models/Platform.swift
@@ -1,0 +1,17 @@
+//
+//  Platform.swift
+//  VideoGameDatabase
+//
+//  Created by macbook on 12/8/25.
+//
+import Foundation
+
+struct PlatformItem: Codable {
+    let platform: Platform
+}
+
+struct Platform: Codable {
+    let id: Int
+    let name: String
+    let slug: String
+}
