+++ VideoGameDatabase/VideoGameDatabase/Services/NetworkingClient.swift
@@ -7,13 +7,22 @@
 
 import Foundation
 
-class NetworkingClient {
+final class NetworkingClient {
     static let shared = NetworkingClient()
     
     private let apiKey = "32f88c2bfb2c4abab74d0914c61b78e7"
     private let baseURL = "https://api.rawg.io/api"
     
+    
+    private let decoder: JSONDecoder = {
+        let decoder = JSONDecoder()
+        decoder.keyDecodingStrategy = .convertFromSnakeCase
+        return decoder
+    }()
+    
     private init() {}
+    
+    // MARK: - Public Methods
     
     func fetch<T: Codable>(endpoint: String, completion: @escaping (Result<T, APIError>) -> Void) {
         
@@ -37,29 +46,30 @@ class NetworkingClient {
                     completion(.failure(.requestFailed(error)))
                     return
                 }
+
+                guard let httpResponse = response as? HTTPURLResponse else {
+                    completion(.failure(.invalidData))
+                    return
+                }
                 
-                guard let httpResponse = response as? HTTPURLResponse, (200...299).contains(httpResponse.statusCode) else {
-                    completion(.failure(.invalidResponse))
+                guard (200...299).contains(httpResponse.statusCode) else {
+                    completion(.failure(.serverError(statusCode: httpResponse.statusCode)))
                     return
                 }
                 
                 guard let data = data else {
-                    completion(.failure(.invalidResponse))
+                    completion(.failure(.invalidData))
                     return
                 }
                 
                 do {
-                    let decoder = JSONDecoder()
-                    decoder.keyDecodingStrategy = .convertFromSnakeCase
-                    
-                    let decodedObject = try decoder.decode(T.self, from: data)
+                    let decodedObject = try self.decoder.decode(T.self, from: data)
                     completion(.success(decodedObject))
                 } catch {
                     completion(.failure(.decodingError(error)))
                 }
             }
         }
-        
         task.resume()
     }
 }
